thought of a slicker way to do this, without so much disruption
