+++ python/tank/commands/clone_configuration.py
@@ -12,7 +12,8 @@ from ..util import ShotgunPath
 from ..errors import TankError
 from . import constants
 from ..util import filesystem
-from ..util import yaml_load, yaml_safe_dump
+
+from tank_vendor import yaml
 
 from .action_base import Action
 
@@ -230,7 +231,7 @@ def _do_clone(log, tk, source_pc_id, user_id, new_name, target_linux, target_mac
         # read the file first
         fh = open(sg_pc_location, "rt")
         try:
-            data = yaml_load(fh)
+            data = yaml.load(fh)
         finally:
             fh.close()
 
@@ -260,7 +261,7 @@ def _do_clone(log, tk, source_pc_id, user_id, new_name, target_linux, target_mac
         # >>> yaml.safe_dump({"foo": u"bar"})
         # '{foo: bar}\n'
         #
-        yaml_safe_dump(data, fh)
+        yaml.safe_dump(data, fh)
         fh.close()
 
     except Exception, e:
