+++ python/tank/util/__init__.py
@@ -34,6 +34,3 @@ from . import filesystem
 from .local_file_storage import LocalFileStorageManager
 
 from .errors import UnresolvableCoreConfigurationError, ShotgunAttachmentDownloadError, EnvironmentVariableFileLookupError
-
-from .yaml_commands import yaml_load, yaml_load_preserve
-from .yaml_commands import yaml_dump, yaml_dump_preserve, yaml_safe_dump
